navbar reorganisation for recruitment and dept members
