+++ src/components/department/MemberCard.tsx
@@ -8,19 +8,24 @@ import {
   Text,
 } from '@chakra-ui/react'
 import { Avatar } from 'flowbite-react'
-import DocumentModal from './DocumentModal'
+import DocumentModal from '../recruitment/director/DocumentModal'
 import type { User } from '~/server/db/models/User'
 
 const MemberCard = ({ member }: { member: User }) => {
   return (
     <Card
-      maxWidth="xs"
+      maxWidth="250px"
       borderRadius="20"
       boxShadow="0px 4px 4px rgba(0, 0, 0, 0.25)"
       minHeight="250px"
     >
       <CardHeader paddingBottom="0" zIndex={1}>
-        <Heading size="md" textAlign="center" textColor="white">
+        <Heading
+          size="md"
+          textAlign="center"
+          textColor="white"
+          overflowWrap="break-word"
+        >
           {member.name}
         </Heading>
       </CardHeader>
@@ -36,10 +41,7 @@ const MemberCard = ({ member }: { member: User }) => {
       <CardBody className="relative" pt="10px">
         <Stack>
           <Avatar alt="User-settings" rounded={true} size="lg" />
-          <DocumentModal
-            name={member.name}
-            resume={member.resume}
-          />
+          <DocumentModal name={member.name} resume={member.resume} />
           <Text className="text-center">{member.department.toUpperCase()}</Text>
           <Text>Role: {member.role} </Text>
         </Stack>
